+++ src/main/java/database/ConnectionManager.java
@@ -1,4 +1,4 @@
-package persistence;
+package database;
 
 import java.sql.Connection;
